+++ java-common/src/main/java/org/duckdns/hjow/colonization/elements/city/City.java
@@ -1368,7 +1368,7 @@ public abstract class City implements HasLocation {
     }
     
     /** 해당 위치의 모든 함선들 반환 */
-    public Vector<Ship> getShips(int x, int y, int z) {
+    public Vector<Ship> getShips(long x, long y, long z) {
     	Vector<Ship> list = new Vector<Ship>();
     	for(Ship s : getShips()) {
     		if(s.getX() == x && s.getY() == y && s.getZ() == z) { list.add(s); }
@@ -1377,7 +1377,7 @@ public abstract class City implements HasLocation {
     }
     
     /** 해당 위치의 해당 범위 내 모든 함선들 반환 */
-    public Vector<Ship> getShips(int x, int y, int z, int dist) {
+    public Vector<Ship> getShips(long x, long y, long z, long dist) {
     	Vector<Ship> list = new Vector<Ship>();
     	for(Ship s : getShips()) {
     		if(DataUtil.getDistance(x, y, z, s.getX(), s.getY(), s.getZ()) <= dist) { list.add(s); }
